rename token auth to password auth
