+++ packages/server/src/aggregates/collection/infrastructure/collection.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { CollectionAbstractService } from '@/aggregates/collection/domain';
 import { CollectionController } from '@/aggregates/collection/presentation';
-import { TokenAuthModule } from '@/aggregates/token-auth/infrastructure';
+import { PasswordAuthModule } from '@/aggregates/password-auth/infrastructure';
 import { EntityModule } from '@/core/modules/module-config';
 
 import { CollectionService } from './collection.service';
@@ -10,7 +10,7 @@ import { CollectionUnitOfWork } from './collection.unit-of-work';
 
 @Module(
   EntityModule.config({
-    imports: [TokenAuthModule],
+    imports: [PasswordAuthModule],
     controllers: [CollectionController],
     providers: [CollectionUnitOfWork],
     services: [
