+++ packages/server/src/aggregates/password-auth/domain/index.ts
