+++ packages/server/src/aggregates/password-auth/domain/token-auth.abstract.service.ts
