+++ packages/server/src/aggregates/password-auth/infrastructure/entities/index.ts
