+++ packages/server/src/aggregates/password-auth/infrastructure/index.ts
@@ -1,0 +1,5 @@
+export * from './entities';
+export * from './repositories';
+export * from './password-auth.module';
+export * from './password-auth.service';
+export * from './password-auth.unit-of-work';
