+++ packages/server/src/aggregates/password-auth/presentation/index.ts
@@ -1,0 +1,1 @@
+export * from './password-auth.controller';
