--- packages/server/src/aggregates/token-auth/infrastructure/index.ts
@@ -1,5 +1,0 @@
-export * from './entities';
-export * from './repositories';
-export * from './token-auth.module';
-export * from './token-auth.service';
-export * from './token-auth.unit-of-work';
