--- packages/server/src/aggregates/token-auth/presentation/index.ts
@@ -1,1 +1,0 @@
-export * from './token-auth.controller';
