+++ packages/server/src/app/app.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { CollectionModule } from '@/aggregates/collection/infrastructure';
 import { RatingSystemModule } from '@/aggregates/rating-system/infrastructure';
 import { SessionModule } from '@/aggregates/session/infrastructure';
-import { TokenAuthModule } from '@/aggregates/token-auth/infrastructure';
+import { PasswordAuthModule } from '@/aggregates/password-auth/infrastructure';
 import { AuthModule } from '@/core/modules/auth';
 import { ConfigModule } from '@/core/modules/config';
 import { CookieModule } from '@/core/modules/cookie';
@@ -16,7 +16,7 @@ import { JsonSchemaModule } from '@/core/modules/json-schema';
   imports: [
     // auth
     AuthModule,
-    TokenAuthModule,
+    PasswordAuthModule,
     // tools
     ConfigModule,
     DatabaseModule,
