+++ packages/server/src/core/modules/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Global, Module } from '@nestjs/common';
 import { DataSource } from 'typeorm';
 
-import { TokenAuthModule } from '@/aggregates/token-auth/infrastructure';
+import { PasswordAuthModule } from '@/aggregates/password-auth/infrastructure';
 import { SessionAbstractRepository } from '@/entities/session/domain';
 import { SessionRepository } from '@/entities/session/infrastructure';
 import { UserRepository } from '@/entities/user/infrastructure';
@@ -10,7 +10,7 @@ import { AuthService } from './auth.service';
 
 @Global()
 @Module({
-  imports: [forwardRef(() => TokenAuthModule)],
+  imports: [forwardRef(() => PasswordAuthModule)],
   providers: [
     AuthService,
     {
